Made various XML and JSON objects apply functions package private in hakija. - Bumped version number since this is a backwards compatibility breaking change. - Also moved the objects to their own package - Moved Hakija sukupuoli resolving to its own object to reduce code duplication
